# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Parking Seeker API",
        "description": "API para la gestión de parkings, incluyendo creación, consulta por ID y búsqueda del parking más cercano a una ubicación geográfica. También registra ubicaciones cuando no se encuentra un parking a menos de 500 metros.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/parkings": {
            "get": {
                "tags": [
                    "3. Parkings"
                ],
                "summary": "Listar parkings paginados",
                "operationId": "e18f3872eb7e2c2c29f2b97bb0ad70ae",
                "responses": {
                    "200": {
                        "description": "Lista paginada de parkings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "nombre": {
                                                                "type": "string",
                                                                "example": "Calle Rivadavia y Santiago"
                                                            },
                                                            "direccion": {
                                                                "type": "string",
                                                                "example": "Calle Rivadavia y Santiago"
                                                            },
                                                            "latitud": {
                                                                "type": "string",
                                                                "example": "-26.8248300"
                                                            },
                                                            "longitud": {
                                                                "type": "string",
                                                                "example": "-65.2001600"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2025-07-31T23:17:59.000000Z"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2025-07-31T23:17:59.000000Z"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "first_page_url": {
                                                    "type": "string",
                                                    "example": "http://localhost:8080/api/parkings?page=1"
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "next_page_url": {
                                                    "type": "string",
                                                    "example": "http://localhost:8080/api/parkings?page=2"
                                                },
                                                "path": {
                                                    "type": "string",
                                                    "example": "http://localhost:8080/api/parkings"
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "prev_page_url": {
                                                    "type": "string",
                                                    "example": null
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    },
                    "405": {
                        "$ref": "#/components/responses/MethodNotAllowed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "3. Parkings"
                ],
                "summary": "Crear un nuevo parking",
                "operationId": "2345ce2d01d1cd69e0c3eef1efcd0a1e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "direccion",
                                    "latitud",
                                    "longitud"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Parking Centro"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "maxLength": 500,
                                        "example": "Calle Rivadavia 123"
                                    },
                                    "latitud": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -26.8241
                                    },
                                    "longitud": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -65.2226
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Parking creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Parking creado correctamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "Parking Centro"
                                                },
                                                "direccion": {
                                                    "type": "string",
                                                    "example": "Calle Rivadavia 123"
                                                },
                                                "latitud": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": -26.8241
                                                },
                                                "longitud": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": -65.2226
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-08-01T17:30:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-08-01T17:30:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    },
                    "405": {
                        "$ref": "#/components/responses/MethodNotAllowed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "responses": {
            "Unauthenticated": {
                "description": "No autenticado o token inválido",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "type": "string",
                                    "example": "No autenticado. Token no proporcionado o inválido."
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Recurso no encontrado",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Recurso no encontrado"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ValidationError": {
                "description": "Datos inválidos",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Datos inválidos."
                                },
                                "errors": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ServerError": {
                "description": "Error interno del servidor",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Error interno del servidor"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "MethodNotAllowed": {
                "description": "Método HTTP no permitido",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Método HTTP no permitido para esta ruta."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Introduce tu token JWT aquí. Ejemplo: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}